dpif-dpdk: add/delete bridge ports correctly.
Currently bridge ports are treated as OUT ports only. Unlike IN
ports, OUT ports should be added to each core in the dataplane.
Similarly when deleting a bridge, only OUT ports should be deleted
and not IN ports (there is no IN port component to a bridge port).
This was not correctly implemented. This patch corrects this.

* dpif-dpdk.c: Modify dpif_dpdk_port_add() to check port type
  to avoid setting IN port flag for bridge ports. Bridge ports
  now only added as part of OUT port transactions.
* dpif-dpdk.c : Modify del_port() to check port type before
  transacting to delete IN ports.

Change-Id: Iecaa34a4805fafb4a42e042a138871d6361a9037
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Reviewed-By: Mark Kavanagh <mark.b.kavanagh@intel.com>
Reviewed-by: Ciara Loftus <ciara.loftus@intel.com>